pyproject.toml
[project]
name = "agent-newsletter-summariser-tts"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "ddgs>=9.12.1",
    "lxml>=6.0.2",
    "ollama>=0.6.1",
    "openai>=2.30.0",
    "python-dotenv>=1.2.2",
    "requests>=2.33.1",
    "smolagents>=1.24.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
